back/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.7.0
python-dotenv==1.0.0
openai>=1.12.0
python-multipart==0.0.6
google-api-python-client==2.108.0
redis==5.0.1
celery==5.3.4
flower==2.0.1
# SQLite3 is built into Python, no additional package needed 